pyproject.toml
[tool.poetry]
name = "data-pipeline"
version = "0.1.0"
description = ""
authors = ["James Potter <jameskpotter80@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
psycopg2-binary = "^2.9.9"
openai = "^1.35.7"
python-dotenv = "^1.0.1"
jupyter = "^1.0.0"
matplotlib = "^3.9.0"
wordcloud = "^1.9.3"
seaborn = "^0.13.2"
anthropic = "^0.30.1"
asyncio = "^3.4.3"
aiofiles = "^24.1.0"
asyncpg = "^0.29.0"
numpy = "^2.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
